+++ web/app/page.tsx
@@ -1,9 +1,10 @@
 import { GrantStatusCountBadges } from "@/components/ui/grant-status-count-badges"
 import { getUser } from "@/lib/auth/user"
 import database from "@/lib/database"
+import { isGrantApproved } from "@/lib/database/helpers"
 import { getPool } from "@/lib/database/queries/pool"
 import { Status } from "@/lib/enums"
-import { getEthAddress, isGrantApproved } from "@/lib/utils"
+import { getEthAddress } from "@/lib/utils"
 import { VotingProvider } from "@/lib/voting/voting-context"
 import { Grant } from "@prisma/flows"
 import Link from "next/link"
