+++ web/lib/utils.ts
@@ -1,12 +1,9 @@
-import { Grant } from "@prisma/flows"
 import { type ClassValue, clsx } from "clsx"
 import { twMerge } from "tailwind-merge"
 import { base, baseSepolia, mainnet } from "viem/chains"
 import { nounsTokenAddress } from "./abis"
 import { NOUNS_TOKEN } from "./config"
-import { Status } from "./enums"
 import { getPinataUrl } from "./pinata/get-file-url"
-import { SupportedChainId } from "./wagmi/config"
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs))
@@ -19,7 +16,7 @@ export function getShortEthAddress(address?: string | null) {
 }
 
 export function explorerUrl(address: string, chainId: number, type: "tx" | "address" = "tx") {
-  const explorerDomain: Record<SupportedChainId, string> = {
+  const explorerDomain: Record<number, string> = {
     [mainnet.id]: "etherscan.io",
     [base.id]: "basescan.org",
     [baseSepolia.id]: "sepolia.basescan.org/",
@@ -27,7 +24,7 @@ export function explorerUrl(address: string, chainId: number, type: "tx" | "addr
 
   if (!(chainId in explorerDomain)) throw new Error("Unsupported chain")
 
-  return `https://${explorerDomain[chainId as SupportedChainId]}/${type}/${address}`
+  return `https://${explorerDomain[chainId]}/${type}/${address}`
 }
 
 export function getIpfsUrl(url: string, gateway: "pinata" | "ipfs" = "ipfs") {
@@ -67,22 +64,6 @@ export function openseaNftUrl(contract: string, tokenId: string, chainId: number
   }
 
   return `${url}/${contract}/${tokenId}`
-}
-
-export function isGrantApproved(grant: Grant) {
-  const { status } = grant
-  return status === Status.Registered || status === Status.ClearingRequested
-}
-
-export function isGrantChallenged(grant: Grant) {
-  const { status, isDisputed, isResolved } = grant
-
-  if (status === Status.ClearingRequested) return true
-  return status === Status.RegistrationRequested && isDisputed === 1 && isResolved === 0
-}
-
-export function isGrantAwaiting(grant: Grant) {
-  return grant.status === Status.RegistrationRequested
 }
 
 export function usesTestNounsToken() {
